+++ lib/src/src.dart
@@ -3,8 +3,7 @@ export 'utils/utils.dart'
     hide AnsiColor, LoggerLogsData, FormatLogger, LoggerNetworkSettings;
 export 'widgets/widgets.dart'
     hide
-        LoggerToast,
-        FloatingLoggerRowText,
+        LoggerToast, 
         FloatingLoggerModalBottomWidget,
         FloatingLoggerItem;
 export 'network/network.dart' hide LogRepository, LogRepositoryModel;