+++ lib/src/widgets/widgets.dart
@@ -1,5 +1,4 @@
 export 'floating_logger_item.dart';
 export 'floating_logger_toast.dart';
-export 'floating_logger_widget.dart';
-export 'floating_logger_row_text.dart';
+export 'floating_logger_widget.dart'; 
 export 'floating_logger_show_modal.dart';